Move modified_patchs_matches to util for reuse
